requirements.txt
retrying==1.3.3
defusedxml==0.6.0
requests==2.21.0
discord.py==1.2.3